refactor(optim)!: Move tuning fns into handlers as pure functions
Relocate tune_temperature and tune_cooling_rate from optimizer modules into their handler modules as free functions. Drop delegate-only associated wrappers. Optimizers call shared pure functions directly. No behavior change.

BREAKING CHANGE: Metropolis::tune_temperature, SimulatedAnnealing::tune_cooling_rate and related associated wrappers removed; use optim::tune_temperature and optim::tune_cooling_rate instead
