+++ src/optim/parallel_tempering.rs
@@ -4,10 +4,8 @@ use ordered_float::NotNan;
 use rand::RngExt as _;
 use rayon::prelude::*;
 
-use super::{
-    GenericLocalSearchOptimizer, LocalSearchOptimizer, Metropolis,
-    generic::StepResult,
-    metropolis::{calculate_temperature_from_acceptance_prob, gather_energy_diffs},
+use super::{GenericLocalSearchOptimizer, LocalSearchOptimizer, Metropolis,
+    calculate_temperature_from_acceptance_prob, gather_energy_diffs, generic::StepResult,
 };
 use crate::{
     Duration, Instant, OptModel,
@@ -98,8 +96,10 @@ impl ParallelTemperingOptimizer {
         if energy_diffs.is_empty() {
             return self;
         }
-        let beta_max = calculate_temperature_from_acceptance_prob(&energy_diffs, target_max_prob);
-        let beta_min = calculate_temperature_from_acceptance_prob(&energy_diffs, target_min_prob);
+        let beta_max =
+            calculate_temperature_from_acceptance_prob(&energy_diffs, target_max_prob);
+        let beta_min =
+            calculate_temperature_from_acceptance_prob(&energy_diffs, target_min_prob);
         let n_replicas = self.betas.len();
         Self::with_geometric_betas(
             self.patience,
